Moved model to data package
