+++ app/src/main/java/com/example/deliveryapp/ui/screens/restaurant/RestaurantScreen.kt
@@ -21,7 +21,7 @@ import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.example.deliveryapp.model.Restaurant
+import com.example.deliveryapp.data.model.Restaurant
 import com.example.deliveryapp.ui.components.menu.Menu
 import com.example.deliveryapp.ui.components.restaurant.RestaurantHeader
 import com.example.deliveryapp.ui.components.shared.AppNavigationBar
